MAINT: move solver.py
