+++ smol/capp/generate/groundstate/solver.py
